Update patch set 2
Patch Set 2:

I'm a little worried that this will break crashreporting on Firefox Android builds again, but GCC just really doesn't like this code, apparently.

Our failure mode was that GCC was miscompiling this somehow and the value returned from sys_clone was wrong, which breaks everything. We only discovered this when we stopped compiling the Linux handler code with -fomit-frame-pointer.

Patch-set: 2
CC: Gerrit User 1134731 <1134731@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>