Fix use of non-whitelisted syscalls on Android.
Android O has a default seccomp policy for all app processes which
disallows most syscalls that aren't used by bionic on x86_64, including
the syscalls which were removed in aarch64 but still supported for
legacy apps in x86_64 in the kernel. Expand the polyfills/conditionals
to cover these when __ANDROID__ is defined on x86_64.

Also, use wait4 instead of waitpid on Android for all architectures,
because waitpid is not permitted even on 32-bit ABIs.

BUG=chromium:770449