Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> we include <unistd.h> already which provides getpagesize().  anyone know of a compelling reason we don't do that (and delete the manual extern) ?
> 
> then again, having a sysconf implementation at all seems a bit odd for the LSS project.  there is no such syscall.  should we just delete it entirely ?  i can't find any users in cs/ or cs.chromium.org.

Whatever floats your boat :) In cs/ the definition for the copy in base:: was changed to no-throw because it's no-throw in shm.h, and this copy is transitively referenced through breakpad. I have no idea what this is being used for, but this breaks build when forcing clang's libc.

Patch-set: 2
Reviewer: Gerrit User 1237241 <1237241@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>