Avoid conflict with pread64/pwrite64 defined by libc
musl libc defines these to pread and pwrite respectively, since its
off_t is always 64-bit. This causes unintended macro replacement
when defining the pread64/pwrite64 wrappers.

To prevent this, save the macro definitions at the start of the
header and restore them at the end. This is what is already done
to work around the same issue on Android NDK.

Bug: linux-syscall-support:23
Change-Id: I0a91ba83fb172818a80a3584302f2c97966ec2e0