Add submit requirements for project linux-syscall-support
Rules for creating new submit requirements are as follows:
* We don't create submit requirements for labels that are not blocking.
* For labels that do not define any values, we assume they are
overriding a label definition from the parent project (i.e. to ignore
this label in the child project), hence we create a non-applicable
submit requirement for them.
* For other blocking labels, we create a corresponding submit
requirement for them.

We also convert the blocking label functions to non-blocking, since
submittability rules are now enforced with the newly defined submit
requirements.

Change-Id: I33fd4ed1adedd1e6a067b11949f06b0e1d363cbd